tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.stroke.overlap.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.stroke.overlap</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.stroke.overlap</h1>
<p class="desc">Stroked subpaths are combined before being drawn</p>


<script>
var t = async_test("Stroked subpaths are combined before being drawn");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#000';
  ctx.fillRect(0, 0, 100, 50);

  ctx.strokeStyle = 'rgba(0, 255, 0, 0.5)';
  ctx.lineWidth = 50;
  ctx.moveTo(0, 20);
  ctx.lineTo(100, 20);
  ctx.moveTo(0, 30);
  ctx.lineTo(100, 30);
  ctx.stroke();

  _assertPixelApprox(canvas, 50,25, 0,127,0,255, 1);
  t.done();

});
</script>
